Moving synchronous validation from show() to constructor
Need to make this change to the actual spec as discussed. It is more
intuitive for developers to get a validation error from data they pass
into the constructor from the constructor itself rather than later when
show() is called. We should only postpone validation that requires I/O
and needs to be asynchronous.